requirements.txt
google-adk>=1.25.1
google-cloud-bigquery>=3.25.0
google-genai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
